Disable service config cache for deployd
This makes caching optional on list_all_instances_for_service. This is
to fix a bug I found whilst writing the itests. If a config changes on
disk and adds a new instance deployd won't always spot it because of the
caching in this funciton. As a result it doesn't get bounced.